requirements.txt
Pillow==9.3.0
pypdf==3.6.0